Order Cancellation Changes - Deepakk - 12/02/2025

+++ BackEnd/myapp/asset/models/CustomerMasterModel.py
@@ -48,8 +48,8 @@ class CustomerMaster(models.Model):
 
 
 
-    credit_limit = models.DecimalField(max_digits=10, decimal_places=2,null=True, blank=True)
-    used_limit = models.DecimalField(max_digits=10, decimal_places=2,null=True, blank=True ,default=0)
+    credit_limit = models.FloatField(null=True, blank=True)
+    used_limit = models.FloatField(null=True, blank=True ,default=0)
     credit_days = models.CharField(max_length=50, blank=True, null=True)
     contact_person = models.CharField(max_length=255)
     contact_number = models.CharField(max_length=50)
